Change exit process
On Linux, it seems that dynamically loaded libraries
are unloaded from the process memory upon exit *before*
the atexit() callback gets executed. This is not the
case on MS Windows.

This causes segfault upon exit of cap32 if an IPF file
was loaded: during dsk_eject() (called by atexit()), we
call some functions from the CAPS library, which is not
loaded anymore.

Fixing this by removing the atexit() callback and explicitely
cleaning up before we exit.

Another possibility would have been not to clean up at all:
after all when exit is called, the process is terminated so
all dangling structures will be freed by the OS.