Handle changed shift states on key release to fix remapping
See issue #167 for discussion.

When the shift state is different for a key press and release event
and where keys are remapped to different physical CPC keys, the emulator
can receive key up/downs on different keys which results in 'stuck',
repeating keys. This fix tracks the shift states of physical PC keys to
ensure the correct keys are pressed/released on the emulator.