Remove the CApplication and CView singletons.
A few other cleanup included, like removing the CLog (using Caprice32
logging mechanism instead).

This is a step towards allowing to display multiple UIs simulataneously.
Although this is now technically possible (for example pressing Shift+F2 to
start the DevTools and then F1 to start the regular GUI) it's not
working well yet because the events are not properly dispatched.